disable indexes + triggers on manual import
On init, those are created automatically.
When initing with the data they are created after the data is inserted.
But when initing empty to import already downloaded dumps, the indexes etc. are already created and slow down the import massively.
This commit disables all triggers and index updates before import_main and re-enables them afterwards. the last action is to re-index the complete database.